Implement splitting AsyncPtyMaster into halves
Add an AsyncPtyMaster::split() method that allows splitting the
AsyncPtyMaster into an AsyncPtyMasterReadHalf implementing Read and
AsyncRead, as well an AsyncPtyMasterWriteHalf implementing Write and
AsyncWrite.

This functionality mirrors tokio_io::AsyncRead::split(),
therefore the code is mostly taken from tokio/tokio-io/src/split.rs.

The advantage of doing this over using tokio_io::AsyncRead::split() is
that we can implement own traits over the halves.

Also add a Trait AsAsyncPtyFd that allows asynchronously getting the
RawFd backing the PTY master. This is implemented for AsyncPtyMaster,
as well as for the halves returned by AsyncPtyMaster::split()